allow login with prefix of "<domain>\" for selected domains
Some WebDAV clients force to use a domain name for authentication. That
domain name is prepended to the username and separated from it using a
backslash.

This commit allows any user to login with a prefix of "WORKGROUP\",
"NC\", or "NEXTCLOUD\" added to their username. The respective prefix is
removed from the given username when trying to login, if the login
failed using the original username. The same logic is applied when
testing login tokens.

In particular, this can help allowing access to Nextcloud for devices
with an interface to SharePoint that forces to specify a domain name for
authentication. As an example, HP LaserJet M428/M429 (and many others)
can then "Scan to Nextcloud" by using the "Scan to SharePoint" feature
(wich is essentially WebDav) and configuring a domain name of "NC",
"NEXTCLOUD" or "WORKGROUP".

Signed-off-by: Bernhard Kirchen <schlimmchen@posteo.net>